Export a BandwidthEstimate histogram via prometheus
This is stage 1 of the BBR quic changes, this helps us to
understand how much of a difference the BBR changes are going
to make, since it the current state of my BBR stack, it's
far more aggressive than questionably it should be. That means
that it will send more data than will be rx'd on the other side.

This means that directly metering the bandwidth from the other
side is no longer a 100% valid strategy for measuring success.

Ideally in the future we will find a way to export the acknowledged
bytes metric and use that as a goodput number for clients. And monitor
that closely. On CUBIC the output rate is very close to acked_bytes,
but in BBR that might not be true anymore. Even worse in very
throttled or high random loss situations.